server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Akash NP <akashnp1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
langchain = "^0.0.333"
openai = "0.28.1"
python-decouple = "^3.8"
unstructured = "^0.10.29"
markdown = "^3.5.1"
ruff = "^0.1.5"
pre-commit = "^3.5.0"
pymongo = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
